[DIRMINA-1169] Acceptor: fix unbinding a ServerSocketChannel
Since Java 11, closing a ServerSocketChannel may release the socket
only on the next select() call. If we try to register a new binding for
an unbound socket before the next select, a BindException may be thrown.

Ensure that there is a select() call between unbinding some sockets and
binding new sockets. Mark the unbind futures as "done" only after that
intervening select().